minor code fix
